git-svn-id: svn://svn.silverstripe.com/silverstripe/open/modules/sapphire/trunk@60468 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ filesystem/Filesystem.php
@@ -22,8 +22,10 @@ class Filesystem extends Object {
 	
 	/**
 	 * Remove a directory and all subdirectories and files
+	 * @param $contentsOnly If this is true then the contents of the folder will be removed but not the folder itself
 	 */
-	static function removeFolder($folder) {
+	static function removeFolder( $folder, $contentsOnly = false ) {
+		
 		// remove a file encountered by a recursive call.
 		if( !is_dir( $folder ) || is_link($folder) )
 			unlink( $folder );
@@ -36,7 +38,8 @@ class Filesystem extends Object {
 					self::removeFolder( $folder.'/'.$file );
 			
 			closedir($dir);
-			rmdir($folder);
+			
+			if(!$contentsOnly) rmdir($folder);
 		}
 	}
